:recycle: Split schema libraries into packages
Refs #213
