+++ flama/schemas/_libs/marshmallow/schemas/__init__.py
@@ -1,0 +1,9 @@
+from flama.schemas._libs.marshmallow.schemas import (  # noqa: F401  - registers schemas in the shared SCHEMAS singleton
+    llm_anthropic,
+    llm_native,
+    llm_ollama,
+    llm_openai,
+    ml,
+    pagination,
+)
+from flama.schemas._libs.marshmallow.schemas.core import SCHEMAS as SCHEMAS
